+++ src/Fragments/UI/views/livewire/_partials/add-fragment-selection.blade.php
@@ -1,34 +1,38 @@
 <x-slot name="header">
-    <x-chief::dialog.drawer.header title="Fragment toevoegen">
+    <x-chief::dialog.drawer.header
+        title="Fragment toevoegen"
+        x-data
+        x-on:chieftab.window="if ($event.detail.reference === 'add-fragment-tabs') { $wire.onTabChanged($event.detail.id) }"
+    >
         <x-chief::tabs
             size="base"
-            wire:key="add-fragment-tabs-{{ Str::random() }}"
+            wire:key="add-fragment-tabs-header"
+            reference="add-fragment-tabs"
             active-tab="{{ $this->showExisting() ? 'existing' : 'new' }}"
             :show-tabs="false"
         >
-            <x-chief::tabs.tab wire:key="add-fragment-tab-new-{{ Str::random() }}" tab-id="new" tab-label="Nieuw" />
-            <x-chief::tabs.tab
-                wire:key="add-fragment-tab-existing-{{ Str::random() }}"
-                tab-id="existing"
-                tab-label="Bestaand"
-            />
+            <x-chief::tabs.tab wire:key="add-fragment-tab-header-new" tab-id="new" tab-label="Nieuw" />
+            <x-chief::tabs.tab wire:key="add-fragment-tab-header-existing" tab-id="existing" tab-label="Bestaand" />
         </x-chief::tabs>
     </x-chief::dialog.drawer.header>
 </x-slot>
 
 <x-chief::tabs
     size="base"
-    wire:key="add-fragment-tabs-{{ Str::random() }}"
+    wire:key="add-fragment-tabs-content"
+    reference="add-fragment-tabs"
     active-tab="{{ $this->showExisting() ? 'existing' : 'new' }}"
     :show-nav="false"
     :should-listen-for-external-tab="true"
 >
-    <x-chief::tabs.tab wire:key="add-fragment-tab-new-{{ Str::random() }}" tab-id="new">
+    <x-chief::tabs.tab wire:key="add-fragment-tab-content-new" tab-id="new">
         @include('chief-fragments::livewire._partials.add-fragment-new')
     </x-chief::tabs.tab>
 
-    <x-chief::tabs.tab wire:key="add-fragment-tab-existing-{{ Str::random() }}" tab-id="existing">
-        @include('chief-fragments::livewire._partials.add-fragment-existing')
+    <x-chief::tabs.tab wire:key="add-fragment-tab-content-existing" tab-id="existing">
+        @if ($this->shouldRenderExistingTab())
+            @include('chief-fragments::livewire._partials.add-fragment-existing')
+        @endif
     </x-chief::tabs.tab>
 </x-chief::tabs>